cambio tipo de dato long a int para ID

+++ restApi-apicom/src/main/java/cl/apicom/spring/backend/entities/Lista.java
@@ -18,10 +18,10 @@ public class Lista {
 	
 	@Id
 	@Column(name = "id", unique = true, nullable = false)
-	private long id;
+	private int id;
 	
 	@Column(name = "id_user", unique = false, nullable = false)
-	private long id_user;
+	private int id_user;
 	
 	@Column(name = "price", unique = false, nullable = false)
 	private int price;
@@ -30,15 +30,15 @@ public class Lista {
 	private Timestamp list_date;
 
 	
-	public long getId() {
+	public int getId() {
 		return id;
 	}
 	
-	public long getId_user() {
+	public int getId_user() {
 		return id_user;
 	}
 
-	public void setId_user(long id_user) {
+	public void setId_user(int id_user) {
 		this.id_user = id_user;
 	}
